Fixed anim montages bug on client, fixed bug with unit spawn out of the map
Fixed a bug with anim montages, the fact that on the client they could not be played because of the bug with the network code.
Also fixed a bug when unit which was spawned by the building being on top of the map was spawned out of the map due to the determining spawning location code.